feat(database): Split actions into user and project actions. relavant imports were renamed
