+++ project/app/page.tsx
@@ -1,7 +1,7 @@
 import Link from "next/link"
 import { ArrowRight, CheckCircle, Users, Kanban } from "lucide-react"
 import { ThemeToggle } from "@/components/theme-toggle"
-import { getUserData } from "@/actions/actions"
+import { getUserData } from "@/actions/user_actions"
 import { SignedIn, SignedOut, SignIn, SignInButton, SignUp, SignUpButton, UserButton } from "@clerk/nextjs";
 
 export default function HomePage() {
